+++ adapters/Facebook/CHANGELOG.md
@@ -1,8 +1,19 @@
 ## Facebook iOS Mediation Adapter Changelog
 
 #### Version  6.2.0.0
+- Verified compatibility with FAN SDK 6.2.0.
+
+Built and tested with
 - Google Mobile Ads SDK version 7.62.0.
 - FAN SDK 6.2.0.
+
+#### Version 6.0.0.0
+- Verified compatibility with FAN SDK 6.0.0.
+- Now requires Google Mobile Ads SDK version 7.66.0 or higher.
+
+Built and tested with
+- Google Mobile Ads SDK version 7.66.0.
+- FAN SDK version 6.0.0.
 
 #### Version 5.10.1.0
 - Verified compatibility with FAN SDK 5.10.1.